third_party/blink/web_tests/shadow-dom/gethtml-use-counter.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<body>
<script>
test(() => {
    const kElementGetHTML = 4781; // From web_feature.mojom
    assert_false(internals.isUseCounted(document, kElementGetHTML));
    document.body.getHTML({serializableShadowRoots: true});
    assert_true(internals.isUseCounted(document, kElementGetHTML));
}, 'Use of getHTML is use counted.');
</script>
